Move addElevation to tileset
